+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 
 ### Changed
  - Bumped spec target version to 0.21
+ - Revised HTML block parsing to conform to new spec (jgm/commonmark.js@99bd473)
  - Imposed 9-digit limit on ordered list markers, per spec
  - Allow non-initial hyphens in html tag names (jgm/CommonMark#239)
  - Updated list of block tag names